soln to day 17 parts 1 and 2
This is kind of a cheese solution to be honest. Doesn't work for all
possible "program" inputs for the three-bit machine, just the exact one
that I was provided as the instructions and operands where hand
compiled.

For part 2, the same idea can be applied with a program that parses and
runs any possible program input.